requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
peft>=0.5.0
accelerate>=0.20.0
bitsandbytes>=0.41.0
pillow>=9.0.0
pandas>=1.5.0
numpy>=1.24.0
matplotlib>=3.7.0
pyyaml>=6.0
tqdm>=4.65.0
wandb>=0.15.0
scikit-learn>=1.2.0
scipy>=1.10.0
